api-fintech/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.1
psycopg2-binary==2.9.9
structlog==24.4.0
